modified:   db.sqlite3 	modified:   lab_Management/__pycache__/urls.cpython-310.pyc 	modified:   lab_admin/__pycache__/urls.cpython-310.pyc 	modified:   lab_admin/__pycache__/views.cpython-310.pyc 	modified:   lab_admin/views.py 	modified:   student_page/__pycache__/urls.cpython-310.pyc

+++ lab_admin/views.py
@@ -9,15 +9,18 @@ def problem(request):
     if request.method=="POST":
         id_list=[]
         tickets = ticket.objects.all()
-        last=ticket.objects.last()
-        l_id=last.id
-        for ids in range(1 , l_id+1) :
-            if request.POST.get(str(ids)) != None :
-                id=request.POST.get(str(ids))  
-                id_list.append(id)
-        for ids in id_list:
-            instance = ticket.objects.get(id=ids)
-            instance.delete()    
+        a=tickets.exists()
+        if a : 
+
+            last=ticket.objects.last()
+            l_id=last.id
+            for ids in range(1 , l_id+1) :
+                if request.POST.get(str(ids)) != None :
+                    id=request.POST.get(str(ids))  
+                    id_list.append(id)
+            for ids in id_list:
+                instance = ticket.objects.get(id=ids)
+                instance.delete()    
         return render(request, 'student_query.html', {'tickets': tickets})
     tickets = ticket.objects.all()
     return render(request, 'student_query.html', {'tickets': tickets})
